TIP-221: move catalog exceptions
