+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/AbstractFilter.php
@@ -70,7 +70,7 @@ abstract class AbstractFilter implements FilterInterface
      * @param string|array $operator the operator used to filter
      * @param string|array $value    the value(s) to filter
      *
-     * @throws \Pim\Bundle\CatalogBundle\Exception\ProductQueryException
+     * @throws \Pim\Component\Catalog\Exception\ProductQueryException
      *
      * @return string
      */
@@ -89,7 +89,7 @@ abstract class AbstractFilter implements FilterInterface
      * @param string             $locale    the locale
      * @param string             $scope     the scope
      *
-     * @throws \Pim\Bundle\CatalogBundle\Exception\ProductQueryException
+     * @throws \Pim\Component\Catalog\Exception\ProductQueryException
      *
      * @return string
      */
