+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Sorter/BaseSorter.php
@@ -141,7 +141,7 @@ class BaseSorter implements AttributeSorterInterface, FieldSorterInterface
      * @param string             $locale    the locale
      * @param string             $scope     the scope
      *
-     * @throws \Pim\Bundle\CatalogBundle\Exception\ProductQueryException
+     * @throws \Pim\Component\Catalog\Exception\ProductQueryException
      *
      * @return string
      */
