+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Sorter/EntitySorter.php
@@ -91,7 +91,7 @@ class EntitySorter implements AttributeSorterInterface
      * @param string             $locale    the locale
      * @param string             $scope     the scope
      *
-     * @throws \Pim\Bundle\CatalogBundle\Exception\ProductQueryException
+     * @throws \Pim\Component\Catalog\Exception\ProductQueryException
      *
      * @return string
      */
