+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Sorter/MetricSorter.php
@@ -75,7 +75,7 @@ class MetricSorter implements AttributeSorterInterface
      * @param string             $locale    the locale
      * @param string             $scope     the scope
      *
-     * @throws \Pim\Bundle\CatalogBundle\Exception\ProductQueryException
+     * @throws \Pim\Component\Catalog\Exception\ProductQueryException
      *
      * @return string
      */
