+++ src/Pim/Bundle/EnrichBundle/Filter/ProductValuesEditDataFilter.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\EnrichBundle\Filter;
 
-use Pim\Bundle\CatalogBundle\Exception\ObjectNotFoundException;
+use Pim\Component\Catalog\Exception\ObjectNotFoundException;
 use Pim\Bundle\CatalogBundle\Filter\CollectionFilterInterface;
 use Pim\Bundle\CatalogBundle\Filter\ObjectFilterInterface;
 use Pim\Component\Catalog\Model\AttributeInterface;
